assign1/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
cycler==0.10.0
flake8==3.9.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
